+++ CHANGELOG.md
@@ -54,6 +54,7 @@ Fixes
 - Fixed TextField EnterReceiver not allowing element addition or removal in OnPressed
 - Fixed paragraphs and images with text/image callbacks setting their area dirty excessively when first added
 - Fixed paragraphs with custom style causing an exception when removed from a ui
+- Fixed empty tokenized strings throwing an exception with center or right alignment
 
 Removals
 - Marked RootElement.OnAddedToUi as obsolete in favor of UiSystem.OnRootAdded